Calculate totalCost

+++ src/main/java/com/hareesh/springstatemachine/springstatemachinedemo/payment/domain/Item.java
@@ -3,11 +3,13 @@ package com.hareesh.springstatemachine.springstatemachinedemo.payment.domain;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Item {
 
     @JsonProperty("count")
-    private int count;
+    private BigDecimal count;
 
     private Product product;
 }
